added onload event listener and set it to load when componenet did update, transitioned some of the examples for new setup, changed extract data to accomidate parent node, which disrupts some examples
