+++ src/demo/components/d3Components/ChildComponent.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 var createReactComponents = require('./../../../react-d3/createReactComponents');
+import d3 from 'd3';
 
 
 module.exports = React.createClass({
@@ -26,12 +27,11 @@ module.exports = React.createClass({
       }
 
       if(state[rd3Id]['__data__'] !== null) {
-         console.log('__data__', state[rd3Id]['__data__'])
+         // console.log('__data__', state[rd3Id]['__data__'])
         reactD3Elements[i]['__data__'] = state[rd3Id]['__data__']
       }
 
       if(state[rd3Id]['__on']) {
-        console.log('__on', state[rd3Id]['__on'])
         reactD3Elements[i]['__on'] = state[rd3Id]['__on']
         for(var j = 0; j < state[rd3Id]['__on'].length; j++) {
           reactD3Elements[i].addEventListener(state[rd3Id]['__on'][j]["type"], state[rd3Id]['__on'][j]["listener"], state[rd3Id]['__on'][j]["capture"])
@@ -39,8 +39,24 @@ module.exports = React.createClass({
           
       }
 
+      if(state[rd3Id]['__onmousemove']) {
+        // reactD3Elements[i]['__onmousemove'] = state[rd3Id]['__onmousemove']
+
+        reactD3Elements[i].addEventListener('mousemove', state[rd3Id]['__onmousemove'])
+    
+          
+      }
+
       if(state[rd3Id]['__zoom']) {
         reactD3Elements[i]['__zoom'] = state[rd3Id]['__zoom']
+      }
+
+       if(state[rd3Id]['__onload']) {
+        var callback = state[rd3Id]['__onload']
+        setTimeout(function(){
+          callback();
+        }, 0)
+        
       }
 
       if(state[rd3Id]['__transition__']) {
