+++ src/demo/d3-examples/stackedToBars.js
@@ -1,25 +1,7 @@
 var d3 = require('d3');
 var div = document.createElement('div');
 
-      function bumpLayer(n, o) {
-
-        function bump(a) {
-          var x = 1 / (.1 + Math.random()),
-              y = 2 * Math.random() - .5,
-              z = 10 / (.1 + Math.random());
-          for (var i = 0; i < n; i++) {
-            var w = (i / n - y) * z;
-            a[i] += x * Math.exp(-w * w);
-          }
-        }
-
-        var a = [], i;
-        for (i = 0; i < n; ++i) a[i] = o + o * Math.random();
-        for (i = 0; i < 5; ++i) bump(a);
-        return a.map(function(d, i) { return {x: i, y: Math.max(0, d)}; });
-      }
-
-var n = 4, // number of layers
+ var n = 4, // number of layers
     m = 58, // number of samples per layer
     stack = d3.layout.stack(),
     layers = stack(d3.range(n).map(function() { return bumpLayer(m, .1); })),
@@ -51,9 +33,99 @@ var xAxis = d3.svg.axis()
 var svg = d3.select(div).append("svg")
     .attr("width", width + margin.left + margin.right)
     .attr("height", height + margin.top + margin.bottom)
-    .attr('class', "STB")
   .append("g")
     .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
 
+var layer = svg.selectAll(".layer")
+    .data(layers)
+  .enter().append("g")
+    .attr("class", "layer")
+    .style("fill", function(d, i) { return color(i); });
+
+var rect = layer.selectAll("rect")
+    .data(function(d) { return d; })
+  .enter().append("rect")
+    .attr("x", function(d) { return x(d.x); })
+    .attr("y", height)
+    .attr("width", x.rangeBand())
+    .attr("height", 0);
+
+rect.transition()
+    .delay(function(d, i) { return i * 10; })
+    .attr("y", function(d) { return y(d.y0 + d.y); })
+    .attr("height", function(d) { return y(d.y0) - y(d.y0 + d.y); });
+
+svg.append("g")
+    .attr("class", "x axis")
+    .attr("transform", "translate(0," + height + ")")
+    .call(xAxis);
+
+
+function transitionGrouped() {
+  y.domain([0, yGroupMax]);
+  var rect = d3.selectAll("rect");
+  rect.transition()
+      .duration(500)
+      .delay(function(d, i) { return i * 10; })
+      .attr("x", function(d, i, j) { return x(d.x) + x.rangeBand() / n * j; })
+      .attr("width", x.rangeBand() / n)
+    .transition()
+      .attr("y", function(d) { return y(d.y); })
+      .attr("height", function(d) { return height - y(d.y); });
+}
+
+function transitionStacked() {
+
+  y.domain([0, yStackMax]);
+  var rect = d3.selectAll("rect");
+  rect.transition()
+      .duration(500)
+      .delay(function(d, i) { return i * 10; })
+      .attr("y", function(d) { return y(d.y0 + d.y); })
+      .attr("height", function(d) { return y(d.y0) - y(d.y0 + d.y); })
+    .transition()
+      .attr("x", function(d) { return x(d.x); })
+      .attr("width", x.rangeBand());
+}
+
+// Inspired by Lee Byron's test data generator.
+function bumpLayer(n, o) {
+
+  function bump(a) {
+    var x = 1 / (.1 + Math.random()),
+        y = 2 * Math.random() - .5,
+        z = 10 / (.1 + Math.random());
+    for (var i = 0; i < n; i++) {
+      var w = (i / n - y) * z;
+      a[i] += x * Math.exp(-w * w);
+    }
+  }
+
+  var a = [], i;
+  for (i = 0; i < n; ++i) a[i] = o + o * Math.random();
+  for (i = 0; i < 5; ++i) bump(a);
+  return a.map(function(d, i) { return {x: i, y: Math.max(0, d)}; });
+}
+
+svg.on("load", function(){
+    applyFunctions();
+});
+
+function applyFunctions(){
+	var timeout = setTimeout(function() {
+    	d3.select("input[value=\"grouped\"]").property("checked", true).each(change);
+    }, 2000);
+
+	d3.selectAll("input").on("change", change);
+
+	function change() {
+	    clearTimeout(timeout);
+	    if (this.value === "grouped") transitionGrouped();
+	    else transitionStacked();
+	}
+}
 
 module.exports = div
+
+
+
