+++ utils/applyD3ReactId.js
@@ -14,12 +14,17 @@ function applyD3ReactId(children, counter) {
 
       result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count] = {};
 
-      child['__data__']
+      child.hasOwnProperty('__data__')
         ? result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__data__'] = child['__data__']
         : result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__data__'] = null
 
       if(child['__on']) {
         result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__on'] = child['__on']
+      }
+
+      if(child['__onload']){
+        result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__onload'] = child['__onload']
+
       }
 
       if(child['__zoom']) {
@@ -28,6 +33,10 @@ function applyD3ReactId(children, counter) {
 
       if(child['__transition__']) {
         result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__transition__'] = child['__transition__']
+      }
+
+      if(child['__onmousemove']) {
+        result.state[child.localName + '.' + counter + '.' + parentCount + '.' + count]['__onmousemove'] = child['__onmousemove']
       }
 
 
@@ -42,7 +51,6 @@ function applyD3ReactId(children, counter) {
 
   apply(children);
   result.children = children;
-
   return result;
 }
